requirements.txt
Pygments==1.6rc1
dbkit==0.2.0
web.py==0.37
python-creole==1.0.6
